+++ src/main/java/nextstep/subway/line/domain/Section.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section.domain;
+package nextstep.subway.line.domain;
 
 import nextstep.subway.station.domain.Station;
