+++ src/main/java/nextstep/subway/line/dto/SectionRequest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.section.dto;
+package nextstep.subway.line.dto;
 
 public class SectionRequest {
     private Long upStationId;
