--- src/main/java/nextstep/subway/section/ui/SectionController.java
@@ -1,34 +1,0 @@
-package nextstep.subway.section.ui;
-
-import nextstep.subway.section.application.SectionService;
-import nextstep.subway.section.dto.SectionRequest;
-import org.springframework.dao.DataIntegrityViolationException;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class SectionController {
-
-    private final SectionService sectionService;
-
-    public SectionController(SectionService sectionService) {
-        this.sectionService = sectionService;
-    }
-
-    @PostMapping("/lines/{lineId}/sections")
-    public ResponseEntity<Void> addSection(
-        @PathVariable Long lineId,
-        @RequestBody SectionRequest sectionRequest) {
-        sectionService.addSection(lineId, sectionRequest);
-        return ResponseEntity.ok().build();
-    }
-
-    @ExceptionHandler({DataIntegrityViolationException.class, IllegalArgumentException.class})
-    public ResponseEntity<Void> handleIllegalArgsException() {
-        return ResponseEntity.badRequest().build();
-    }
-}
