+++ src/test/java/nextstep/subway/section/SectionAcceptanceTest.java
@@ -5,7 +5,7 @@ import io.restassured.response.ValidatableResponse;
 import nextstep.subway.AbstractAcceptanceTest;
 import nextstep.subway.line.dto.LineRequest;
 import nextstep.subway.line.dto.LineResponse;
-import nextstep.subway.section.dto.SectionRequest;
+import nextstep.subway.line.dto.SectionRequest;
 import nextstep.subway.station.StationAcceptanceTest;
 import nextstep.subway.station.dto.StationResponse;
 import org.junit.jupiter.api.BeforeEach;
